--- samples/app/src/main/kotlin/com/bumble/appyx/app/node/backstack/app/custombackstack/CustomBackStackElement.kt
@@ -1,5 +1,0 @@
-package com.bumble.appyx.app.node.backstack.app.custombackstack
-
-import com.bumble.appyx.core.navigation.NavElement
-
-typealias CustomBackStackElement<NavTarget> = NavElement<NavTarget, CustomBackStack.State>
