+++ samples/app/src/main/kotlin/com/bumble/appyx/app/node/backstack/app/indexedbackstack/IndexedBackStackElement.kt
@@ -1,0 +1,5 @@
+package com.bumble.appyx.app.node.backstack.app.indexedbackstack
+
+import com.bumble.appyx.core.navigation.NavElement
+
+typealias IndexedBackStackElement<NavTarget> = NavElement<NavTarget, IndexedBackStack.State>
