+++ samples/app/src/main/kotlin/com/bumble/appyx/app/node/backstack/app/indexedbackstack/operation/UpdateSize.kt
@@ -1,17 +1,17 @@
-package com.bumble.appyx.app.node.backstack.app.custombackstack.operation
+package com.bumble.appyx.app.node.backstack.app.indexedbackstack.operation
 
-import com.bumble.appyx.app.node.backstack.app.custombackstack.CustomBackStack
-import com.bumble.appyx.app.node.backstack.app.custombackstack.CustomBackStack.State.Destroyed
-import com.bumble.appyx.app.node.backstack.app.custombackstack.CustomBackStack.State.Stashed
+import com.bumble.appyx.app.node.backstack.app.indexedbackstack.IndexedBackStack
+import com.bumble.appyx.app.node.backstack.app.indexedbackstack.IndexedBackStack.State.Destroyed
+import com.bumble.appyx.app.node.backstack.app.indexedbackstack.IndexedBackStack.State.Stashed
 import com.bumble.appyx.core.navigation.NavElements
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-class UpdateSize<T : Any> : CustomBackStackOperation<T> {
+class UpdateSize<T : Any> : IndexedBackStackOperation<T> {
 
-    override fun isApplicable(elements: NavElements<T, CustomBackStack.State>): Boolean = true
+    override fun isApplicable(elements: NavElements<T, IndexedBackStack.State>): Boolean = true
 
-    override fun invoke(elements: NavElements<T, CustomBackStack.State>): NavElements<T, CustomBackStack.State> {
+    override fun invoke(elements: NavElements<T, IndexedBackStack.State>): NavElements<T, IndexedBackStack.State> {
         val destroyedCount = elements.count { it.targetState is Destroyed }
         return elements.mapIndexed { index, element ->
             when (element.targetState) {
